refactor: change names of cal_coefficients and receiver_calibration_func
