+++ pyproject.toml
@@ -165,7 +165,7 @@ ignore-init-module-imports = true
 "src/edges_cal/spectra.py" = [
     "N802",
 ]
-"src/edges_cal/cal_coefficients.py" = [
+"src/edges_cal/calobs.py" = [
     "N802",
 ]
 "src/edges_cal/reflection_coefficient.py" = [
